# api/infra/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/category/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all categories.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Retrieve a list of categories",
                "operationId": "list-categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of categories to be retrieved",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination of categories",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/categorydtos.CategoryOutputDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/validation_errors.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/maintenance/latest/{carID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get latest maintenance associated with a specific carID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get latest maintenance by carID",
                "operationId": "get-latest-maintenances-by-carID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CarID",
                        "name": "carID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved maintenances",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maintenancedtos.MaintenanceOutputDTO"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validation_errors.HTTPError"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/maintenance/maintenance/by-date-range": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get maintenance records within the specified date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get maintenance records by date range",
                "operationId": "get-maintenances-by-date-range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date of the range (format: '2006-01-02')",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date of the range (format: '2006-01-02')",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved maintenance records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maintenancedtos.MaintenanceOutputDTO"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validation_errors.HTTPErrorCar"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/maintenance/{carID}/maintenances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of maintenances associated with a specific carID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get maintenances by carID",
                "operationId": "get-maintenances-by-carID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CarID",
                        "name": "carID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved maintenances",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maintenancedtos.MaintenanceOutputDTO"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validation_errors.HTTPError"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/maintenance/{maintenanceID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a maintenance with the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenance"
                ],
                "summary": "Delete a maintenance",
                "operationId": "delete-maintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maintenance ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintenance deleted successfully"
                    },
                    "400": {
                        "description": "Error details",
                        "schema": {
                            "$ref": "#/definitions/validation_errors.HTTPErrorCar"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing maintenance with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Maintenance"
                ],
                "summary": "Update an existing maintenance",
                "operationId": "patch-maintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "maintenanceID",
                        "name": "maintenanceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Maintenance information to be updated",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/maintenancedtos.MaintenanceInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated maintenance",
                        "schema": {
                            "$ref": "#/definitions/maintenancedtos.MaintenanceOutputDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "categorydtos.CategoryOutputDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "enums.MaintenanceStatus": {
            "type": "string",
            "enum": [
                "Scheduled",
                "In Progress",
                "Completed",
                "Pending Approval",
                "Canceled",
                "Awaiting Parts",
                "Awaiting Payment",
                "Rescheduled",
                "Maintenance Failed",
                "Awaiting Inspection"
            ],
            "x-enum-varnames": [
                "Scheduled",
                "InProgress",
                "Completed",
                "PendingApproval",
                "Canceled",
                "AwaitingParts",
                "AwaitingPayment",
                "Rescheduled",
                "MaintenanceFailed",
                "AwaitingInspection"
            ]
        },
        "maintenancedtos.MaintenanceInputDTO": {
            "type": "object",
            "required": [
                "last_maintenance_date",
                "maintenance_completion_date",
                "maintenance_status",
                "maintenance_type",
                "next_maintenance_due_date",
                "odometer_reading"
            ],
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "labor_cost": {
                    "type": "integer",
                    "minimum": 1
                },
                "last_maintenance_date": {
                    "type": "string"
                },
                "maintenance_completion_date": {
                    "type": "string"
                },
                "maintenance_duration": {
                    "type": "string"
                },
                "maintenance_notes": {
                    "type": "string"
                },
                "maintenance_status": {
                    "$ref": "#/definitions/enums.MaintenanceStatus"
                },
                "maintenance_type": {
                    "type": "string"
                },
                "next_maintenance_due_date": {
                    "type": "string"
                },
                "odometer_reading": {
                    "type": "integer",
                    "minimum": 0
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maintenancedtos.PartInputDTO"
                    }
                },
                "parts_cost": {
                    "type": "integer",
                    "minimum": 1
                },
                "scheduled_maintenance": {
                    "type": "boolean"
                }
            }
        },
        "maintenancedtos.MaintenanceOutputDTO": {
            "type": "object",
            "required": [
                "last_maintenance_date",
                "maintenance_completion_date",
                "maintenance_status",
                "next_maintenance_due_date"
            ],
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labor_cost": {
                    "type": "integer"
                },
                "last_maintenance_date": {
                    "type": "string"
                },
                "maintenance_completion_date": {
                    "type": "string"
                },
                "maintenance_duration": {
                    "type": "string"
                },
                "maintenance_notes": {
                    "type": "string"
                },
                "maintenance_status": {
                    "$ref": "#/definitions/enums.MaintenanceStatus"
                },
                "maintenance_type": {
                    "type": "string"
                },
                "next_maintenance_due_date": {
                    "type": "string"
                },
                "odometer_reading": {
                    "type": "integer"
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maintenancedtos.PartOutputDTO"
                    }
                },
                "parts_cost": {
                    "type": "integer"
                },
                "scheduled_maintenance": {
                    "type": "boolean"
                }
            }
        },
        "maintenancedtos.PartInputDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "cost": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "replacement_date": {
                    "type": "string"
                }
            }
        },
        "maintenancedtos.PartOutputDTO": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "maintenance_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "replacement_date": {
                    "type": "string"
                }
            }
        },
...
        "validation_errors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "validation_errors.HTTPErrorCar": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 422
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['name is required'",
                        " 'description is required'",
                        " 'licenseplate is required'",
                        " 'brand is required'",
                        " 'A car needs a category to be registered']"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}